refactor: 필요없는 인터페이스 삭제
