--- src/main/java/co/fineants/api/domain/notification/domain/entity/policy/NotificationEvaluator.java
@@ -1,5 +1,0 @@
-package co.fineants.api.domain.notification.domain.entity.policy;
-
-public interface NotificationEvaluator<T> {
-	boolean isSatisfiedBy(T target);
-}
