+++ src/main/java/co/fineants/api/domain/notification/domain/entity/policy/TargetGainNotificationPolicy.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.notification.domain.entity.policy.target_gain;
+package co.fineants.api.domain.notification.domain.entity.policy;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,14 +6,13 @@ import java.util.function.Predicate;
 
 import co.fineants.api.domain.common.notification.Notifiable;
 import co.fineants.api.domain.notification.domain.dto.response.NotifyMessage;
-import co.fineants.api.domain.notification.domain.entity.policy.NotificationPolicy;
 import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
 import co.fineants.api.domain.portfolio.domain.entity.Portfolio;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class TargetGainNotificationPolicy implements NotificationPolicy<Notifiable> {
-	
+
 	private final List<Predicate<Portfolio>> portfolioConditions;
 	private final Predicate<NotificationPreference> preferencePredicate;
