+++ src/main/java/co/fineants/api/domain/notification/domain/entity/policy/TargetPriceNotificationPolicy.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.notification.domain.entity.policy.target_price;
+package co.fineants.api.domain.notification.domain.entity.policy;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,6 @@ import java.util.function.Predicate;
 
 import co.fineants.api.domain.common.notification.Notifiable;
 import co.fineants.api.domain.notification.domain.dto.response.NotifyMessage;
-import co.fineants.api.domain.notification.domain.entity.policy.NotificationPolicy;
 import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
 import co.fineants.api.domain.stock_target_price.domain.entity.TargetPriceNotification;
 import lombok.RequiredArgsConstructor;
@@ -15,15 +14,14 @@ import lombok.RequiredArgsConstructor;
 public class TargetPriceNotificationPolicy implements NotificationPolicy<Notifiable> {
 
 	private final List<Predicate<TargetPriceNotification>> targetPriceConditions;
-	private final List<Predicate<NotificationPreference>> preferenceConditions;
+	private final Predicate<NotificationPreference> preferenceConditions;
 
 	@Override
 	public Optional<NotifyMessage> apply(Notifiable target, String token) {
-		boolean result = targetPriceConditions.stream()
-			.allMatch(condition -> condition.test((TargetPriceNotification)target))
-			&& preferenceConditions.stream()
-			.allMatch(condition -> condition.test(target.getNotificationPreference()));
-		if (result) {
+		boolean isTargetPriceValid = targetPriceConditions.stream()
+			.allMatch(condition -> condition.test((TargetPriceNotification)target));
+		boolean isPreferenceValid = preferenceConditions.test(target.getNotificationPreference());
+		if (isTargetPriceValid && isPreferenceValid) {
 			return Optional.of(target.getTargetPriceMessage(token));
 		}
 		return Optional.empty();
