+++ src/test/java/co/fineants/api/domain/notification/service/provider/FirebaseNotificationProviderTest.java
@@ -27,8 +27,8 @@ import co.fineants.api.domain.kis.repository.CurrentPriceRedisRepository;
 import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.domain.member.repository.MemberRepository;
 import co.fineants.api.domain.notification.domain.dto.response.SentNotifyMessage;
-import co.fineants.api.domain.notification.domain.entity.policy.maxloss.MaxLossNotificationPolicy;
-import co.fineants.api.domain.notification.domain.entity.policy.target_gain.TargetGainNotificationPolicy;
+import co.fineants.api.domain.notification.domain.entity.policy.MaxLossNotificationPolicy;
+import co.fineants.api.domain.notification.domain.entity.policy.TargetGainNotificationPolicy;
 import co.fineants.api.domain.portfolio.domain.entity.Portfolio;
 import co.fineants.api.domain.portfolio.repository.PortfolioRepository;
 import co.fineants.api.domain.purchasehistory.domain.entity.PurchaseHistory;
